Move some shared code into the base controller
Make method shared
