+++ app/controllers/legal_services/sessions_controller.rb
@@ -3,7 +3,6 @@ module LegalServices
     protected
 
     def challenge_path
-      cookies[:session] = { value: @result.session, expires: 20.minutes, httponly: true }
       legal_services_users_challenge_path(challenge_name: @result.challenge_name, username: @result.cognito_uuid)
     end
