+++ app/controllers/legal_services/users_controller.rb
@@ -3,7 +3,6 @@ module LegalServices
     private
 
     def new_challenge_path
-      cookies[:session] = @challenge.new_session
       legal_services_users_challenge_path(challenge_name: @challenge.new_challenge_name, username: params[:username])
     end
