+++ app/controllers/management_consultancy/admin/sessions_controller.rb
@@ -4,7 +4,6 @@ module ManagementConsultancy
       protected
 
       def challenge_path
-        cookies[:session] = { value: @result.session, expires: 20.minutes, httponly: true }
         management_consultancy_admin_users_challenge_path(challenge_name: @result.challenge_name, username: @result.cognito_uuid)
       end
