+++ app/controllers/management_consultancy/admin/users_controller.rb
@@ -4,7 +4,6 @@ module ManagementConsultancy
       private
 
       def new_challenge_path
-        cookies[:session] = @challenge.new_session
         management_consultancy_admin_users_challenge_path(challenge_name: @challenge.new_challenge_name, username: params[:username])
       end
