+++ app/controllers/management_consultancy/sessions_controller.rb
@@ -3,7 +3,6 @@ module ManagementConsultancy
     protected
 
     def challenge_path
-      cookies[:session] = { value: @result.session, expires: 20.minutes, httponly: true }
       management_consultancy_users_challenge_path(challenge_name: @result.challenge_name, username: @result.cognito_uuid)
     end
