+++ app/controllers/supply_teachers/admin/sessions_controller.rb
@@ -4,7 +4,6 @@ module SupplyTeachers
       protected
 
       def challenge_path
-        cookies[:session] = { value: @result.session, expires: 20.minutes, httponly: true }
         supply_teachers_admin_users_challenge_path(challenge_name: @result.challenge_name, username: @result.cognito_uuid)
       end
