+++ app/controllers/supply_teachers/users_controller.rb
@@ -3,7 +3,6 @@ module SupplyTeachers
     private
 
     def new_challenge_path
-      cookies[:session] = @challenge.new_session
       supply_teachers_users_challenge_path(challenge_name: @challenge.new_challenge_name, username: params[:username])
     end
