+++ app/helpers/management_consultancy/suppliers_helper.rb
@@ -5,13 +5,6 @@ module ManagementConsultancy::SuppliersHelper
     "#{framework}.lot_#{lot_number}"
   end
 
-  def framework_lot_and_description(number, description)
-    framework = number.split('.')[0].gsub('1', '')
-    lot_number = number.split('.')[1]
-
-    "#{framework} lot #{lot_number} - #{description}"
-  end
-
   def url_formatter(url)
     u = URI.parse(url)
